use size_t instead of int. fix pointer math for free() call. tab indent 2 spaces in js to match other code

+++ src/manifold/include/manifold.h
@@ -98,8 +98,8 @@ class Manifold {
    */
   ///@{
   Mesh GetMesh() const;
-  int GetMeshDirect(int) const;
-  void FreeMeshDirect(int) const;
+  size_t GetMeshDirect(int) const;
+  void FreeMeshDirect(size_t) const;
   bool IsEmpty() const;
   enum class Error {
     NO_ERROR,
